requirements.txt
numpy==1.21.2
pandas==1.3.3
matplotlib==3.4.3
seaborn==0.11.2
scipy==1.7.1
plotly==5.3.1
bokeh==2.3.3
dash==2.0.0
dash-renderer==1.9.1
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-bootstrap-components==0.13.0
ggplot==0.11.5
scikit-learn==0.24.2
statsmodels==0.12.2
jupyter==1.0.0
jupyterlab==3.1.11
ipython==7.27.0
RPy2==3.4.5
psycopg2==2.9.1
pymysql==1.0.2
pymongo==3.12.0
cassandra-driver==3.25.0
apache-airflow==2.1.4
prefect==0.15.4
sqlalchemy==1.4.23
boto3==1.18.65
google-cloud-storage==1.42.0
azure-storage-blob==12.9.0
dask==2021.9.1
hdfs==2.6.0
pyspark==3.1.2
docker==5.0.2
kubernetes==18.20.0
Flask==2.0.1  
Werkzeug==2.0.1  
